# [#2288] public body page performance