--- a/app/controllers/public_body_controller.rb
+++ b/app/controllers/public_body_controller.rb
@@ -82,7 +82,7 @@ class PublicBodyController < ApplicationController
                 @existing_track = TrackThing.find_existing(@user, @track_thing)
             end
 
-            @follower_count = TrackThing.count(:all, :conditions => ["public_body_id = ?", @public_body.id])
+            @follower_count = TrackThing.where(:public_body_id => @public_body.id).count
 
             @feed_autodetect = [ { :url => do_track_url(@track_thing, 'feed'),
                                    :title => @track_thing.params[:title_in_rss],