--- a/app/models/info_request_event.rb
+++ b/app/models/info_request_event.rb
@@ -327,15 +327,15 @@ class InfoRequestEvent < ActiveRecord::Base
 
 
     def is_incoming_message?
-        !incoming_message.nil?
+        incoming_message_id? or (incoming_message if new_record?)
     end
 
     def is_outgoing_message?
-        !self.outgoing_message.nil?
+        outgoing_message_id? or (outgoing_message if new_record?)
     end
 
     def is_comment?
-        !self.comment.nil?
+        comment_id? or (comment if new_record?)
     end
 
     # Display version of status