--- a/app/views/public_body/show.html.erb
+++ b/app/views/public_body/show.html.erb
@@ -57,18 +57,19 @@
         <div class="action-bar__make-request">
             <% if @public_body.is_requestable? || @public_body.not_requestable_reason == 'bad_contact' %>
                 <%= link_to _("Make a request to this authority"), new_request_to_body_path(:url_name => @public_body.url_name), :class => "link_button_green" %>
-           <% end %>
+            <% end %>
         </div>
+
         <div class="action-bar__follow">
-              <div class="action-bar__follow-button">
-              <% if @existing_track %>
-                <%= (link_to _("Unsubscribe"), {:controller => 'track', :action => 'update', :track_id => @existing_track.id, :track_medium => "delete", :r => request.fullpath}, :class => "link_button_green") %>
-              <% else %>
-                <div class="feed_link">
-                      <%= link_to _("Follow"), do_track_path(@track_thing), :class => "link_button_green" %>
-                </div>
-              <% end %>
-              </div>
+            <div class="action-bar__follow-button">
+                <% if @existing_track %>
+                    <%= link_to _("Unsubscribe"), {:controller => 'track', :action => 'update', :track_id => @existing_track.id, :track_medium => "delete", :r => request.fullpath}, :class => "link_button_green" %>
+                <% else %>
+                    <div class="feed_link">
+                        <%= link_to _("Follow"), do_track_path(@track_thing), :class => "link_button_green" %>
+                    </div>
+                <% end %>
+            </div>
 
             <div class="action-bar__follower-count">
                 <%= n_("{{count}} follower",