--- a/app/views/public_body/show.html.erb
+++ b/app/views/public_body/show.html.erb
@@ -136,8 +136,14 @@
         <% if @number_of_visible_requests > 4 %>
             <%= render :partial => 'request/request_search_form' %>
         <% end %>
+
         <%= render :partial => 'more_info', :locals => { :public_body => @public_body } %>
-        <%= render :partial => 'track/tracking_links', :locals => { :track_thing => @track_thing, :own_request => false, :location => 'sidebar' } %>
+
+        <%= render :partial => 'track/tracking_links',
+                   :locals => { :track_thing => @track_thing,
+                                :existing_track => @existing_track,
+                                :own_request => false,
+                                :location => 'sidebar' } %>
     </div>
 
 </div>