--- a/app/views/request/_request_search_form.html.erb
+++ b/app/views/request/_request_search_form.html.erb
@@ -2,22 +2,22 @@
 
 <div id="list-filter">
   <%= form_tag(request.path, :method => "get", :id=>"filter_requests_form") do %>
- <div class="list-filter-item">
-  <%= label_tag(:query, _("Keywords"), :class=>"form_label title") %>
-  <%= text_field_tag(:query, params[:query]) %> 
- </div>
+    <div class="list-filter-item">
+      <%= label_tag(:query, _("Keywords"), :class=>"form_label title") %>
+      <%= text_field_tag(:query, params[:query]) %>
+    </div>
 
- <div class="list-filter-item">
-   <%= label_tag(:query, _("Made between"), :class=>"form_label title") %>
-   <%= text_field_tag(:request_date_after, params[:request_date_after], {:class => "use-datepicker", :size => 10}) %>&nbsp;&nbsp;
-   <%= label_tag(:query, _("and"), :class=>"form_label") %>
-   <%= text_field_tag(:request_date_before, params[:request_date_before], {:class => "use-datepicker", :size => 10}) %>
- </div>
+    <div class="list-filter-item">
+      <%= label_tag(:query, _("Made between"), :class=>"form_label title") %>
+      <%= text_field_tag(:request_date_after, params[:request_date_after], {:class => "use-datepicker", :size => 10}) %>&nbsp;&nbsp;
+      <%= label_tag(:query, _("and"), :class=>"form_label") %>
+      <%= text_field_tag(:request_date_before, params[:request_date_before], {:class => "use-datepicker", :size => 10}) %>
+    </div>
 
-  <%= after_form_fields if defined?(after_form_fields) -%>
+    <%= after_form_fields if defined?(after_form_fields) -%>
 
- <div class="list-filter-item">
-  <%= submit_tag(_("Search")) %>
- </div>
-<% end %>
+    <div class="list-filter-item">
+      <%= submit_tag(_("Search")) %>
+    </div>
+  <% end %>
 </div>