--- a/app/views/request/_request_search_form.html.erb
+++ b/app/views/request/_request_search_form.html.erb
@@ -6,19 +6,7 @@
   <%= label_tag(:query, _("Keywords"), :class=>"form_label title") %>
   <%= text_field_tag(:query, params[:query]) %> 
  </div>
-<% if false # don't think we want this, but leaving as an example %>
- <div class="list-filter-item">
-  <%= _("Search for words in:") %> <br/>
-  <% [["sent", _("messages from users")], 
-     ["response", _("messages from authorities")],
-     ["comment", _("comments")]].each_with_index do |item, index|
-     variety, title = item %>
-  
-   <%= check_box_tag "request_variety[]", variety, params[:request_variety].nil? ? true : params[:request_variety].include?(variety), :id => "request_variety_#{index}" %>
-   <%= label_tag("request_variety_#{index}", title) %> <br/>
-  <% end %>
- </div>
-<% end %>
+
  <div class="list-filter-item">
    <%= label_tag(:query, _("Made between"), :class=>"form_label title") %>
    <%= text_field_tag(:request_date_after, params[:request_date_after], {:class => "use-datepicker", :size => 10}) %>&nbsp;&nbsp;