#2288 

Could only get minor improvements so far:

```
Old: Completed 200 OK in 2966.7ms (Views: 2637.3ms | ActiveRecord: 88.2ms)
New: Completed 200 OK in 2675.2ms (Views: 2514.1ms | ActiveRecord: 94.8ms)
```

The bulk of the response time happens when rendering each search result:

```diff
-                <%= render :partial => 'request/request_listing_via_event', :locals => { :event => result[:model] } %>
+                <%# render :partial => 'request/request_listing_via_event', :locals => { :event => result[:model] } %>
```

```
Completed 200 OK in 711.1ms (Views: 511.3ms | ActiveRecord: 8.3ms)
```

I'd like to release at this point and try to get some more detailed data from New Relic to see where to go next. I think the main issue is that we're [loading the whole result set](https://github.com/mysociety/alaveteli/blob/develop/app/controllers/application_controller.rb#L394-L395) after pulling them out from Xapian. We're then calling out to many association objects during the view rendering. Looks like we can pass an [`:eager_load` option](https://github.com/mysociety/alaveteli/blob/develop/lib/acts_as_xapian/acts_as_xapian.rb#L424) to the Xapian search, but we don't currently support passing options to the [search wrapper method](https://github.com/mysociety/alaveteli/blob/develop/app/controllers/application_controller.rb#L376).