@garethrees 2015-06-04
Merged in https://github.com/mysociety/alaveteli/commit/26dbf38319f22de55d6c8bfbf3376df0dbb297c2