@garethrees 2015-07-07
Also looks to have helped here:

![screen shot 2015-07-07 at 13 30 52](https://cloud.githubusercontent.com/assets/282788/8546073/74b70c10-24ac-11e5-82d1-797bb02588e4.png)

Looks like much less time in `request/_request_listing_via_event.html.erb`.